@@ -1,3 +1,3 @@
-class Gdc < ActiveRecord::Base
+class Gdc < ApplicationRecord
     self.table_name = 'global_daily_cumulative'
 end
