@@ -1,3 +1,3 @@
-class Gdd < ActiveRecord::Base
+class Gdd < ApplicationRecord
     self.table_name = 'global_daily_delta'
 end
