@@ -1,3 +1,3 @@
-class Idc < ActiveRecord::Base
+class Idc < ApplicationRecord
     self.table_name = 'india_daily_cumulative'
 end
