@@ -1,3 +1,3 @@
-class Idd < ActiveRecord::Base
+class Idd < ApplicationRecord
     self.table_name = 'india_daily_delta'
 end
